requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.0; python_version >= '3.7'
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11; python_version >= '3'
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django-allauth==0.47.0
django-cors-headers==3.11.0
django-rest-auth==0.9.5
django==3.2.10
djangorestframework-gis==0.18
djangorestframework==3.13.1
idna==3.3; python_version >= '3'
itypes==1.2.0
jinja2==3.0.3; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
oauthlib==3.2.0; python_version >= '3.6'
psycopg2-binary==2.9.3
pycparser==2.21
pyjwt[crypto]==2.3.0; python_version >= '3.6'
python3-openid==3.2.0
pytz==2021.3
pyyaml==6.0
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.27.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==1.26.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
